fix/PYOKEMON-113-테넌트 서비스, 컨트롤러 분리
